my_server_app/app/api/routers/auth.py
# This file contains API endpoints related to user authentication,
# including login, token generation, and user registration.

from fastapi import APIRouter, Depends, HTTPException, status
from fastapi.security import OAuth2PasswordBearer, OAuth2PasswordRequestForm
from jose import JWTError, jwt
from datetime import timedelta

# Add other necessary imports for your authentication logic (e.g., database interaction)

router = APIRouter()

# Example basic route (you'll need to implement the actual logic)
@router.post("/token")
async def login_for_access_token(form_data: OAuth2PasswordRequestForm = Depends()):
    # Implement your user authentication logic here
    # If authentication is successful, create and return a JWT token
    return {"access_token": "your_jwt_token", "token_type": "bearer"}

# Add other authentication-related endpoints as needed (e.g., /register)