my_server_app/app/database/connection.py
# This file is responsible for managing database connections using SQLAlchemy.

from sqlalchemy import create_engine
from sqlalchemy.ext.declarative import declarative_base
from sqlalchemy.orm import sessionmaker
import psycopg2 # Assuming PostgreSQL is used, include psycopg2 import

# Database connection URL (replace with your actual database URL)
DATABASE_URL = "postgresql://user:password@host:port/dbname"

# Create a SQLAlchemy engine
engine = create_engine(DATABASE_URL)

# Create a configured "Session" class
SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)

# Create a Base class for declarative models
Base = declarative_base()

def get_db():
    """Dependency to get a database session."""
    db = SessionLocal()
    try:
        yield db
    finally:
        db.close()